renaming the directories
